IGNITE-5112 Create OLS example:   Created example & tests refactoring.
